Add close method to escrow contract
If a campaign is fully funded the creator of the campaign can close
the contract. This will cause a new DAO with the campaign's governance
token to be instantiated.

Notably, this will not transfer the campaign's funds or minting
control to the DAO. A followup commit will add a transfer method for
handling that.